*Jwt tuned *User, Faculty and Department CRUD

+++ src/main/java/com/shabab/UniversityManagementSystem/academy/service/FacultyService.java
@@ -2,10 +2,15 @@ package com.shabab.UniversityManagementSystem.academy.service;
 
 import com.shabab.UniversityManagementSystem.academy.model.Faculty;
 import com.shabab.UniversityManagementSystem.academy.repository.FacultyRepository;
+import com.shabab.UniversityManagementSystem.admin.model.University;
+import com.shabab.UniversityManagementSystem.admin.model.User;
+import com.shabab.UniversityManagementSystem.security.jwt.JwtUtil;
+import com.shabab.UniversityManagementSystem.util.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * Project: UniversityManagementSystem-SpringBoot
@@ -19,20 +24,83 @@ public class FacultyService {
     @Autowired
     private FacultyRepository facultyRepository;
 
-    public void saveFaculty(Faculty faculty) {
-        facultyRepository.save(faculty);
+    @Autowired
+    private JwtUtil jwtUtil;
+
+    public ApiResponse getAll() {
+        ApiResponse response = new ApiResponse();
+        try {
+            List<Faculty> faculties = facultyRepository.findByUniversity(jwtUtil.getUniversity());
+            if (faculties.isEmpty()) {
+                return response.returnError("No faculty found");
+            }
+            response.setData("faculties", faculties);
+            response.success("Successfully retrieved all faculties");
+            return response;
+        } catch (Exception e) {
+            return response.returnError(e);
+        }
     }
 
-    public List<Faculty> getAllFaculties() {
-        return facultyRepository.findAll();
+    public ApiResponse save(Faculty faculty) {
+        ApiResponse response = new ApiResponse();
+        try {
+            faculty.setUniversity(jwtUtil.getUniversity());
+            Faculty dbFaculty = facultyRepository.save(faculty);
+            response.setData("faculty", dbFaculty);
+            response.success("Saved Successfully");
+            return response;
+        } catch (Exception e) {
+            return response.returnError(e);
+        }
     }
 
-    public void deleteFaculty(Long id) {
-        facultyRepository.deleteById(id);
+    public ApiResponse update(Faculty faculty) {
+        ApiResponse response = new ApiResponse();
+        try {
+            Faculty dbFaculty = facultyRepository.findByIdAndUniversity(faculty.getId(), jwtUtil.getUniversity());
+            if (dbFaculty == null || dbFaculty.getId() == null) {
+                return response.returnError("Faculty not found");
+            }
+
+            faculty.setUniversity(jwtUtil.getUniversity());
+            Faculty updatedFaculty = facultyRepository.save(faculty);
+            response.setData("faculty", updatedFaculty);
+            response.success("Updated Successfully");
+            return response;
+        } catch (Exception e) {
+            return response.returnError(e);
+        }
     }
 
-    public Faculty getFacultyById(Long id) {
-        return facultyRepository.findById(id).get();
+    public ApiResponse getById(Long id) {
+        ApiResponse response = new ApiResponse();
+        try {
+            Faculty faculty = facultyRepository.findByIdAndUniversity(id, jwtUtil.getUniversity());
+            if (faculty == null || faculty.getId() == null) {
+                return response.returnError("Faculty not found");
+            }
+            response.setData("faculty", faculty);
+            response.success("Successfully retrieved faculty");
+            return response;
+        } catch (Exception e) {
+            return response.returnError(e);
+        }
+    }
+
+    public ApiResponse deleteById(Long id) {
+        ApiResponse response = new ApiResponse();
+        try {
+            Faculty faculty = facultyRepository.findByIdAndUniversity(id, jwtUtil.getUniversity());
+            if (faculty == null || faculty.getId() == null) {
+                return response.returnError("Faculty not found");
+            }
+            facultyRepository.deleteById(id);
+            response.success("Deleted Successfully");
+            return response;
+        } catch (Exception e) {
+            return response.returnError(e);
+        }
     }
 
 }
